Splitting concepts responsibility
Separated the concepts and responsibilities of the Code Generators to avoid the confusion caused by partial classes.

I applied RawString to format the unit manager code. This makes the code more readable and eliminates the use of AppendLineLF.

I reorganized the namespaces to focus efforts on the code generators and avoid the distractions of utility classes. All the business logic is now under Generators.

Utilities and satellite classes are under Tools.

Extension methods under Extensions.

Diagnostics have been grouped under the Rules class. I added a new one to track fatal errors and list it in the output.

I also added various extensions to streamline the code.

Reworking the code, I noticed that by not using the CodeGeneration option. The different classes don't go through the CurrentUICulture (which is problematic if you're using the tool in a Roslyn project).

Now that the class responsibilities are clearly defined, I could add a new Generator to the IStringLocalization model. This interface is used in WinUI and Blazor projects, among others.