Added compatibility with IStringLocalizer
Using heritage pattern over the generators to clarify witch every class do. To clarify and simplify concept, separation of concepts.

Unification of all Diagnostics in a class Analyzer to respect the Analyzer contract from MS RoslynAnalyser.

Rebrand the SourceGenerator class to switch over a new option GenerationType :
  - ResourceManager (Default)
  - CodeGeneration (In addidition to GenerateCode, to maintenability of existing code)
  - StringLocaliser (new generator includes registration classes to simplifing adding resources to the CollectionService) I added a demo project to demonstrate the new feature.

I optimized the uses of IncrementalProviders to avoid to apply the same filter over and over.

Now we have thee bloc of generation
- ResourcesManager
- CodeGeneration
   - ComboGeneration
- StringLocaliser
  - NamespaceRegistrator
  - GlobalRegistrator