Correction of somes glitch after using it in real environement
The global registration class generated even if no StringLocalizer was present
Pass the message of exception instead of exception in case of fatal error
Environement.NewLine not recommended for Analyser and Generator. I changed for a constant
Adjust the Unshipped.md
Add a QUICKSTART.md to have a global view of possibilities the option and simplify configuration for new and current users.